Refactor: Add command, query package (#3)
