--- StudentSubscriptionService/src/main/java/com/example/studentsubscriptionservice/application/in/CreateStudentCommand.java
@@ -1,7 +1,0 @@
-package com.example.studentsubscriptionservice.application.in;
-
-public record CreateStudentCommand(
-
-        String name
-) {
-}
