+++ StudentSubscriptionService/src/main/java/com/example/studentsubscriptionservice/application/in/command/CreateStudentCommand.java
@@ -1,0 +1,7 @@
+package com.example.studentsubscriptionservice.application.in.command;
+
+public record CreateStudentCommand(
+
+        String name
+) {
+}
