+++ StudentSubscriptionService/src/main/java/com/example/studentsubscriptionservice/application/usecase/CreateStudentUseCase.java
@@ -1,7 +1,7 @@
 package com.example.studentsubscriptionservice.application.usecase;
 
-import com.example.studentsubscriptionservice.application.in.CreateStudentCommand;
-import com.example.studentsubscriptionservice.application.in.StudentDto;
+import com.example.studentsubscriptionservice.application.in.command.CreateStudentCommand;
+import com.example.studentsubscriptionservice.application.in.dto.StudentDto;
 
 public interface CreateStudentUseCase {
